backend/foodgram/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.0
django-filter==23.5
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
reportlab==4.0.9